requirements.txt
aniso8601==8.0.0
apispec==3.0.0
astroid==2.3.1
attrs==19.2.0
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
flask-restplus==0.13.0
gunicorn==19.9.0
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
marshmallow==3.2.1
mccabe==0.6.1
more-itertools==7.2.0
Pillow==6.2.0
pylint==2.4.2
pyrsistent==0.15.4
pytz==2019.3
six==1.12.0
typed-ast==1.4.0
webargs==5.5.2
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
